+++ niryo_robot_utils/src/niryo_robot_utils/sentry.py
@@ -1,7 +1,10 @@
 import os
 
-import sentry_sdk
-from sentry_sdk import set_tags
+try:
+    import sentry_sdk
+except ImportError:
+    from unittest.mock import MagicMock
+    sentry_sdk = MagicMock()
 
 
 def sentry_init():
@@ -9,12 +12,12 @@ def sentry_init():
     Initialize sentry
     '''
 
-    try:
-        sentry_dsn_value = os.getenv['SENTRY_DSN_PYTHON']
-        sentry_raspid = os.getenv['SENTRY_RASPID']
+    sentry_dsn_value = os.getenv('SENTRY_DSN_PYTHON')
+    sentry_raspid = os.getenv('SENTRY_RASPID')
 
+    try:
         if sentry_dsn_value is not None and sentry_raspid is not None:
-            set_tags({"raspid": str(sentry_raspid)})
+            sentry_sdk.set_tags({"raspid": str(sentry_raspid)})
             sentry_sdk.init(dsn=sentry_dsn_value)
 
     except Exception: